# pkg/apiserver/docs/swagger.yaml
definitions:
...
  api.DrbdEnableSetting:
    properties:
      enable:
        type: boolean
      state:
        $ref: '#/definitions/api.State'
      version:
        type: string
    type: object
  api.DrbdEnableSettingReqBody:
    properties:
      enable:
        type: boolean
    type: object
  api.DrbdEnableSettingRspBody:
    properties:
      data:
        $ref: '#/definitions/api.DrbdEnableSetting'
    type: object
  api.LocalDiskInfo:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      availableCapacityBytes:
        description: AvailableCapacityBytes 可用容量
        type: integer
      diskPathShort:
        description: diskPathShort 磁盘路径简写
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      localStoragePooLName:
        description: LocalStoragePooLName 存储池名称
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1alpha1.LocalDiskSpec'
      status:
        $ref: '#/definitions/v1alpha1.LocalDiskStatus'
      totalCapacityBytes:
        description: TotalCapacityBytes 总容量
        type: integer
    type: object
...
  api.NodeDiskListByPool:
    properties:
      items:
        description: localDisks 节点磁盘列表
        items:
          $ref: '#/definitions/api.LocalDiskInfo'
        type: array
      nodeName:
        description: nodeName 节点名称
        type: string
      pagination:
        allOf:
        - $ref: '#/definitions/api.Pagination'
        description: page 信息
      poolName:
        description: PoolName 存储池名称
        type: string
    type: object
...
  api.Pagination:
    properties:
      page:
        description: 当前页索引，从 1 开始，为 0 时，会自动重置为默认值 constants.DefaultPage
        type: integer
      pageSize:
        description: |-
          每页数据量，为 -1 时表示查询全部，为 0 时会重置为默认值
          constants.DefaultPageSize
        type: integer
      pages:
        description: 总页数
        type: integer
      total:
        description: 总共有多少条目，请求时可以不用传递
        type: integer
    type: object
  api.RspFailBody:
    properties:
      description:
        type: string
      errcode:
        type: integer
    type: object
  api.State:
    enum:
    - Submitted
    - AddReplica
    - SyncReplica
    - PruneReplica
    - InProgress
    - Completed
    - ToBeAborted
    - Cancelled
    - Aborted
    - Failed
    - ToBeMounted
    - ""
    - Created
    - Creating
    - Ready
    - NotReady
    - ToBeDeleted
    - Deleted
    - Unknown
    - Invalid
    - Creating
    - Ready
    - NotReady
    - ToBeDeleted
    - Deleted
    - Ready
    - Maintain
    - Offline
    - ""
    - Unknown
    - ReadyAndNotReady
    - Healthy
    - NotReady
    - ""
    - Ready
    - Maintain
    - Offline
    - Unknown
    - Unclaimed
    - Released
    - Claimed
    - Inuse
    - Reserved
    - ReleaseReserved
    - ""
    - ClaimedAndUnclaimed
    - Pending
    - Bound
    - Available
    - Active
    - Inactive
    - Unknown
    - Running
    - NotReady
    - Enabled
    - Disabled
    - Succeed
    - Failed
    type: string
    x-enum-varnames:
    - OperationStateSubmitted
    - OperationStateMigrateAddReplica
    - OperationStateMigrateSyncReplica
    - OperationStateMigratePruneReplica
    - OperationStateInProgress
    - OperationStateCompleted
    - OperationStateToBeAborted
    - OperationStateAborting
    - OperationStateAborted
    - OperationStateFailed
    - VolumeStateToBeUnmount
    - VolumeStateEmpty
    - VolumeStateCreated
    - VolumeStateCreating
    - VolumeStateReady
    - VolumeStateNotReady
    - VolumeStateToBeDeleted
    - VolumeStateDeleted
    - VolumeStateUnknown
    - VolumeReplicaStateInvalid
    - VolumeReplicaStateCreating
    - VolumeReplicaStateReady
    - VolumeReplicaStateNotReady
    - VolumeReplicaStateToBeDeleted
    - VolumeReplicaStateDeleted
    - NodeStateReady
    - NodeStateMaintain
    - NodeStateOffline
    - NodeStateEmpty
    - NodeStateUnknown
    - NodeStateReadyAndNotReady
    - NodeStateHealthy
    - NodeStateNotReady
    - DriverStateEmpty
    - DriverStateReady
    - DriverStateMaintain
    - DriverStateOffline
    - DriverStateUnknown
    - LocalDiskUnclaimed
    - LocalDiskReleased
    - LocalDiskClaimed
    - LocalDiskInUse
    - LocalDiskReserved
    - LocalDiskReleaseReserved
    - LocalDiskEmpty
    - LocalDiskClaimedAndUnclaimed
    - LocalDiskPending
    - LocalDiskBound
    - LocalDiskAvailable
    - LocalDiskActive
    - LocalDiskInactive
    - LocalDiskUnknown
    - ModuleStatusRunning
    - ModuleStatusNotReady
    - DrbdModuleStatusEnabled
    - DrbdModuleStatusDisabled
    - ReservedSucceedState
    - ReservedFailedState
  api.StorageNode:
    properties:
      K8sNode:
        $ref: '#/definitions/v1.Node'
      k8SNodeState:
        $ref: '#/definitions/api.State'
      localDiskNode:
        $ref: '#/definitions/v1alpha1.LocalDiskNode'
      localStorageNode:
        $ref: '#/definitions/v1alpha1.LocalStorageNode'
    type: object
...
  api.StorageNodeListByPool:
    properties:
      items:
        description: StorageNodes
        items:
          $ref: '#/definitions/api.StorageNode'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/api.Pagination'
        description: page 信息
      storagePoolName:
        description: StoragePoolName 存储池名称
        type: string
    type: object
  api.StorageNodePool:
    properties:
      class:
        description: |-
          Supported class: HDD, SSD, NVMe
          +kubebuilder:validation:Enum:=HDD;SSD;NVMe
        type: string
      disks:
        items:
          $ref: '#/definitions/v1alpha1.LocalDevice'
        type: array
      freeCapacityBytes:
        type: integer
      freeVolumeCount:
        type: integer
      name:
        description: 'Supported pool name: HDD_POOL, SSD_POOL, NVMe_POOL'
        type: string
      nodeName:
        description: NodesName Pool所在节点
        type: string
      path:
        description: VG path
        type: string
      totalCapacityBytes:
        type: integer
      totalVolumeCount:
        type: integer
      type:
        description: |-
          Supported type: REGULAR
          +kubebuilder:validation:Enum:=REGULAR
          +kubebuilder:default:=REGULAR
        type: string
      usedCapacityBytes:
        type: integer
      usedVolumeCount:
        type: integer
      volumeCapacityBytesLimit:
        type: integer
      volumes:
        items:
          type: string
        type: array
    type: object
  api.StoragePool:
    properties:
      allocatedCapacityBytes:
        description: AllocatedCapacityBytes 存储池已经分配存储容量
        type: integer
      createTime:
        description: createTime 创建时间
        type: string
      items:
        items:
          $ref: '#/definitions/api.StorageNodePool'
        type: array
      nodeNames:
        description: NodesNames Pool所在节点列表
        items:
          type: string
        type: array
      poolName:
        description: 'Supported pool name: HDD_POOL, SSD_POOL, NVMe_POOL 存储池名称'
        type: string
      totalCapacityBytes:
        description: TotalCapacityBytes 存储池对应存储总容量
        type: integer
    type: object
  api.StoragePoolExpansionReqBody:
    properties:
      diskType:
        description: HDD/SSD/NVME
        type: string
      nodeName:
        type: string
      owner:
        description: local-storage/local-disk-manager
        type: string
    type: object
  api.StoragePoolExpansionRspBody:
    properties:
      success:
        type: boolean
    type: object
  api.StoragePoolList:
    properties:
      items:
        description: storagePools
        items:
          $ref: '#/definitions/api.StoragePool'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/api.Pagination'
        description: page 信息
    type: object
...
  v1.Node:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        allOf:
        - $ref: '#/definitions/v1.ObjectMeta'
        description: |-
          Standard object's metadata.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
      spec:
        allOf:
        - $ref: '#/definitions/v1.NodeSpec'
        description: |-
          Spec defines the behavior of a node.
          https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
          +optional
      status:
        allOf:
        - $ref: '#/definitions/v1.NodeStatus'
        description: |-
          Most recently observed status of the node.
          Populated by the system.
          Read-only.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
          +optional
    type: object
...
  v1.NodeConfigSource:
    properties:
      configMap:
        allOf:
        - $ref: '#/definitions/v1.ConfigMapNodeConfigSource'
        description: ConfigMap is a reference to a Node's ConfigMap
    type: object
...
  v1.NodeSpec:
    properties:
      configSource:
        allOf:
        - $ref: '#/definitions/v1.NodeConfigSource'
        description: |-
          If specified, the source to get node configuration from
          The DynamicKubeletConfig feature gate must be enabled for the Kubelet to use this field
          +optional
      externalID:
        description: |-
          Deprecated. Not all kubelets will set this field. Remove field after 1.13.
          see: https://issues.k8s.io/61966
          +optional
        type: string
      podCIDR:
        description: |-
          PodCIDR represents the pod IP range assigned to the node.
          +optional
        type: string
      podCIDRs:
        description: |-
          podCIDRs represents the IP ranges assigned to the node for usage by Pods on that node. If this
          field is specified, the 0th entry must match the podCIDR field. It may contain at most 1 value for
          each of IPv4 and IPv6.
          +optional
          +patchStrategy=merge
        items:
          type: string
        type: array
      providerID:
        description: |-
          ID of the node assigned by the cloud provider in the format: <ProviderName>://<ProviderSpecificNodeID>
          +optional
        type: string
      taints:
        description: |-
          If specified, the node's taints.
          +optional
        items:
          $ref: '#/definitions/v1.Taint'
        type: array
      unschedulable:
        description: |-
          Unschedulable controls node schedulability of new pods. By default, node is schedulable.
          More info: https://kubernetes.io/docs/concepts/nodes/node/#manual-node-administration
          +optional
        type: boolean
    type: object
...
  v1alpha1.LocalDevice:
    properties:
      capacityBytes:
        description: disk capacity
        type: integer
      devPath:
        description: e.g. /dev/sdb
        type: string
      state:
        allOf:
        - $ref: '#/definitions/v1alpha1.State'
        description: 'Possible state: Available, Inuse, Offline'
      type:
        description: 'Supported: HDD, SSD, NVMe, RAM'
        type: string
    type: object
...
  v1alpha1.LocalDiskNode:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1alpha1.LocalDiskNodeSpec'
      status:
        $ref: '#/definitions/v1alpha1.LocalDiskNodeStatus'
    type: object
  v1alpha1.LocalDiskNodeSpec:
    properties:
      nodeName:
        description: NodeName represent where disk is attached
        type: string
    type: object
...
  v1alpha1.LocalDiskSpec:
    properties:
      capacity:
        description: Capacity of the disk in bytes
        type: integer
      claimRef:
        allOf:
        - $ref: '#/definitions/v1.ObjectReference'
        description: |-
          ClaimRef is the reference to the LDC which has claimed this LD
          +optional
      devLinks:
        description: DevLinks are symbol links for this device
        items:
          type: string
        type: array
      devicePath:
        description: DevicePath is the disk path in the OS
        type: string
      diskAttributes:
        allOf:
        - $ref: '#/definitions/v1alpha1.DiskAttributes'
        description: DiskAttributes has hardware/static attributes of the disk
      isRaid:
        description: HasRAID identifies if the disk is a raid disk or not
        type: boolean
      major:
        description: |-
          Major represents drive used by the device
          +optional
        type: string
      minor:
        description: |-
          Minor is used to distinguish different devices
          +optional
        type: string
      nodeName:
        description: |-
          NodeName represents the node where the disk is attached
          +kubebuilder:validation:Required
        type: string
      owner:
        description: |-
          Owner represents which system owns this claim(e.g. local-storage, local-disk-manager)
          +optional
        type: string
      partitionInfo:
        description: |-
          PartitionInfo contains partition information
          +optional
        items:
          $ref: '#/definitions/v1alpha1.PartitionInfo'
        type: array
      partitioned:
        description: HasPartition represents if the disk has partitions or not
        type: boolean
      preDevicePath:
        description: PreDevicePath represents the last device path in the OS
        type: string
      preNodeName:
        description: PreNodeName represents the node where the disk was attached
        type: string
      raidInfo:
        allOf:
        - $ref: '#/definitions/v1alpha1.RAIDInfo'
        description: |-
          RAIDInfo contains RAID information
          +optional
      reserved:
        description: |-
          Reserved represents the disk won't be used in hwameistor later, until it becomes unreserved
          +optional
        type: boolean
      smartInfo:
        allOf:
        - $ref: '#/definitions/v1alpha1.SmartInfo'
        description: |-
          SmartInfo contains infos collected by smartctl
          +optional
      state:
        allOf:
        - $ref: '#/definitions/v1alpha1.LocalDiskState'
        description: |-
          State is the current state of the disk (Active/Inactive/Unknown)
          +kubebuilder:validation:Enum:=Active;Inactive;Unknown
      supportSmart:
        description: HasSmartInfo identified if the disk supports SMART or not
        type: boolean
      uuid:
        description: UUID global unique identifier of the disk
        type: string
    type: object
...
  v1alpha1.LocalDiskStatus:
    properties:
      claimState:
        allOf:
        - $ref: '#/definitions/v1alpha1.LocalDiskState'
        description: |-
          State represents the claim state of the disk
          +kubebuilder:validation:Enum:=Bound;Reserved;Available;Pending
    type: object
...
  v1alpha1.LocalStorageNode:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1alpha1.LocalStorageNodeSpec'
      status:
        $ref: '#/definitions/v1alpha1.LocalStorageNodeStatus'
    type: object
  v1alpha1.LocalStorageNodeSpec:
    properties:
      hostname:
        type: string
      storageIP:
        description: IPv4 address is for HA replication traffic
        type: string
      topogoly:
        $ref: '#/definitions/v1alpha1.Topology'
    type: object
...
  v1alpha1.RAIDInfo:
    properties:
      raidMaster:
        description: RAIDMaster is the master of the RAID disk, it works for only
          RAID slave disk, e.g. /dev/bus/0
        type: string
    type: object
...
  v1alpha1.State:
    enum:
    - Ready
    - Maintain
    - Offline
    - ToBeMounted
    - ""
    - Created
    - Creating
    - Ready
    - NotReady
    - ToBeDeleted
    - Deleted
    - Invalid
    - Creating
    - Ready
    - NotReady
    - ToBeDeleted
    - Deleted
    - Consistent
    - Inconsistent
    - Up
    - Down
    - Submitted
    - AddReplica
    - SyncReplica
    - PruneReplica
    - InProgress
    - Completed
    - ToBeAborted
    - Cancelled
    - Aborted
    - Failed
    - Available
    - InUse
    - Offline
    - ""
    - ToBeMounted
    - ToBeUnMount
    - Mounted
    - NotReady
    type: string
    x-enum-varnames:
    - NodeStateReady
    - NodeStateMaintain
    - NodeStateOffline
    - VolumeStateToBeUnmount
    - VolumeStateEmpty
    - VolumeStateCreated
    - VolumeStateCreating
    - VolumeStateReady
    - VolumeStateNotReady
    - VolumeStateToBeDeleted
    - VolumeStateDeleted
    - VolumeReplicaStateInvalid
    - VolumeReplicaStateCreating
    - VolumeReplicaStateReady
    - VolumeReplicaStateNotReady
    - VolumeReplicaStateToBeDeleted
    - VolumeReplicaStateDeleted
    - HAVolumeReplicaStateConsistent
    - HAVolumeReplicaStateInconsistent
    - HAVolumeReplicaStateUp
    - HAVolumeReplicaStateDown
    - OperationStateSubmitted
    - OperationStateMigrateAddReplica
    - OperationStateMigrateSyncReplica
    - OperationStateMigratePruneReplica
    - OperationStateInProgress
    - OperationStateCompleted
    - OperationStateToBeAborted
    - OperationStateAborting
    - OperationStateAborted
    - OperationStateFailed
    - DiskStateAvailable
    - DiskStateInUse
    - DiskStateOffline
    - MountPointStateEmpty
    - MountPointToBeMounted
    - MountPointToBeUnMount
    - MountPointMounted
    - MountPointNotReady
...
info:
  contact: {}
paths:
...
  /cluster/drbd:
    get:
      consumes:
      - application/json
      description: get DRBDSettingGet
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DrbdEnableSetting'
      summary: 摘要 获取高可用设置
      tags:
      - Setting
    post:
      consumes:
      - application/json
      description: post EnableDRBDSetting
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.DrbdEnableSettingReqBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DrbdEnableSettingRspBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 高可用设置
      tags:
      - Setting
...
  /cluster/pools:
    get:
      consumes:
      - application/json
      description: list StoragePools
      parameters:
      - description: name
        in: query
        name: name
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      - description: fuzzy
        in: query
        name: fuzzy
        type: boolean
      - description: sort
        in: query
        name: sort
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StoragePoolList'
      summary: 摘要 获取存储池列表信息
      tags:
      - Pool
  /cluster/pools/{poolName}:
    get:
      consumes:
      - application/json
      description: get Pool angel
      parameters:
      - description: poolName
        in: path
        name: poolName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StoragePool'
      summary: 摘要 获取指定存储池基本信息
      tags:
      - Pool
  /cluster/pools/{poolName}/nodes:
    get:
      consumes:
      - application/json
      description: get StorageNodesGetByPoolName
      parameters:
      - description: poolName
        in: path
        name: poolName
        required: true
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      - description: nodeName
        in: query
        name: nodeName
        type: string
      - description: state
        in: query
        name: state
        type: string
      - description: fuzzy
        in: query
        name: fuzzy
        type: boolean
      - description: sort
        in: query
        name: sort
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StorageNodeListByPool'
      summary: 摘要 获取指定存储池存储节点列表信息
      tags:
      - Pool
  /cluster/pools/{poolName}/nodes/{nodeName}/disks:
    get:
      consumes:
      - application/json
      description: get StorageNodeDisksGetByPoolName
      parameters:
      - description: nodeName
        in: path
        name: nodeName
        required: true
        type: string
      - description: poolName
        in: path
        name: poolName
        required: true
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      - description: fuzzy
        in: query
        name: fuzzy
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.NodeDiskListByPool'
      summary: 摘要 获取指定存储池指定存储节点磁盘列表信息
      tags:
      - Pool
  /cluster/pools/expand:
    post:
      consumes:
      - application/json
      description: expand new disk for storage pool
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.StoragePoolExpansionReqBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StoragePoolExpansionRspBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: Storage pool expand
      tags:
      - Pool
...
swagger: "2.0"